rename entities to mysql
